+++ codex/models/admin.py
@@ -94,7 +94,7 @@ class ComicboxTaggingDefaults(BaseModel):
         NEVER = "never", _("Never")
 
     default_formats = JSONField(default=list)
-    delete_original = BooleanField(default=False)
+    delete_original = BooleanField(default=True)
     default_match_mode = CharField(
         max_length=MAX_FIELD_LEN,
         choices=MatchModeChoices.choices,